requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
django-otp==1.5.4
django-password-validators==1.7.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
filelock==3.16.1
flake8==7.1.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
jwcrypto==1.5.6
mccabe==0.7.0
mongoengine==0.29.1
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
psycopg==3.2.3
psycopg2==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pymongo==4.10.1
pytest==8.3.3
pytest-django==4.9.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
